+++ dantro/plot/utils/_file_writer.py
@@ -5,6 +5,7 @@ import os
 
 import matplotlib as mpl
 import matplotlib.animation
+import matplotlib.figure
 
 # -----------------------------------------------------------------------------